+++ docs/patterns/README.md
@@ -1,0 +1,9 @@
+# Patterns
+
+## Execute a pattern file
+
+While `bash` working directory is the root of EOSFactory, do, for example:
+
+```md
+eosfactory/utils/pythonmd.sh  docs/patterns/set/set_account_permission.md
+```
