+++ docs/patterns/set/set_account_permission.md
@@ -1,10 +1,9 @@
-'''
 # set account permission
 
 Creates or updates an account's permission.
 
-```md
-'''
+```python
+
 from eosfactory.eosf import *
 import eosfactory.core.cleos as cleos
 import eosfactory.core.setup as setup
@@ -14,27 +13,27 @@ create_master_account("master")
 create_account("Alice", master)
 create_account("Jim", master)
 Alice.info()
-'''
+
 ```
-In the *permissions* section, `Alice.info()` reads:
+`Alice.info()` reads, in the *permissions* section:
 ```md
 permissions:
-     owner     1:    1 EOS7Br6tgiwxMU7an4QUKWfkcgRPW1bE4s4AEGP78LMBDyhx9VBa5
-        active     1:    1 EOS5Cv8oujEUZu9Vx3JTeyMzgQEQg64g8mVQrPMyBk6P16eTSxwfg
+     owner     1:    1 Alice@owner
+        active     1:    1 Alice@active
 ```
 
 ## Set new key to a permission
-```md
-'''
+```python
+
 COMMENT("Set new key to a permission:")
 key = cleos.CreateKey(is_verbose=False)
 setup.is_print_command_line = True
-Alice.set_permission(
+Alice.set_account_permission(
     Permission.ACTIVE, key.key_public, Permission.OWNER, 
     (Alice, Permission.OWNER))
 setup.is_print_command_line = False
 Alice.info()
-'''
+
 ```
 Now, the *permissions* section of `Alice.info()` changes:
 ```md
@@ -44,17 +43,17 @@ permissions:
 ```
 
 ## Set an account (instead of a key) as authority for a permission
-```md
-'''
+```python
+
 COMMENT("Set an account (instead of a key) as authority for a permission:")
 create_account("Bob", master)
 setup.is_print_command_line = True
-Alice.set_permission(
+Alice.set_account_permission(
     Permission.ACTIVE, Bob, Permission.OWNER, 
     (Alice, Permission.OWNER))
 setup.is_print_command_line = False
 Alice.info()
-'''
+
 ```
 
 The permissions section of `Alice.info()`:
@@ -69,13 +68,13 @@ permissions:
 
 Note that actors have to be sorted in the ``authority`` JSON. (Why ``cleos`` cannot sort them itself?)
 
-```md
-'''
+```python
+
 COMMENT("Weights and Threshold:")
 create_account("Carol", master)
 actors = [str(Bob), str(Carol)]
 actors.sort()
-Alice.set_permission(Permission.ACTIVE,
+Alice.set_account_permission(Permission.ACTIVE,
     {
         "threshold" : 100, 
         "keys" : [], 
@@ -102,7 +101,7 @@ Alice.set_permission(Permission.ACTIVE,
     Permission.OWNER,
     (Alice, Permission.OWNER))
 Alice.info()
-'''
+
 ```
 The permissions section of `Alice.info()`:
 ```md
@@ -114,13 +113,13 @@ permissions:
 ## Set two weighted keys
 
 Note that keys have to be sorted in the ``authority`` JSON. (Why ``cleos`` cannot sort them itself?)
-```md
-'''
+```python
+
 COMMENT("Set two weighted keys:")
 keys = [Bob.owner(), Carol.owner()]
 keys.sort()
 
-Alice.set_permission(Permission.ACTIVE,
+Alice.set_account_permission(Permission.ACTIVE,
     {
         "threshold" : 100, 
         "keys" : 
@@ -139,7 +138,7 @@ Alice.set_permission(Permission.ACTIVE,
     (Alice, Permission.OWNER)
 )
 Alice.info()
-'''
+
 ```
 The permissions section of `Alice.info()`:
 
@@ -148,4 +147,4 @@ permissions:
      owner     1:    1 Alice@owner
         active   100:    50 Alice@active, 50 Bob@owner
 ```
-'''
+
