+++ eosfactory/shell/account.py
@@ -16,10 +16,15 @@ if setup.node_api == "cleos":
     import eosfactory.core.cleos as cleos
 elif setup.node_api == "eosjs":
     import eosfactory.core.eosjs as cleos
+
+import eosfactory.core.cleos_set as cleos_set
 import eosfactory.core.cleosys as cleosys
 import eosfactory.core.manager as manager
 import eosfactory.core.testnet as testnet
+import eosfactory.core.account_set as account_set
+
 import eosfactory.shell.wallet as wallet
+
 
 '''The namespace where account objects go.
 '''
@@ -55,27 +60,7 @@ def is_local_testnet_running(account_eosio):
     try: # remote eosio may have the ["keys"] array empty.
         return account_eosio.owner_key.key_public == account_.owner_key
     except:
-        False        
-
-
-def _data_json(data):
-    class Encoder(json.JSONEncoder):
-        def default(self, o):
-            if isinstance(o, interface.Account):
-                return str(o)
-            else:
-                json.JSONEncoder.default(self, o) 
-    if not data:
-        return data
-
-    data_json = data
-    if isinstance(data, dict) or isinstance(data, list):
-        data_json = json.dumps(data, cls=Encoder)
-    else:
-        if isinstance(data, str):
-            data_json = re.sub("\s+|\n+|\t+", " ", data)
-            data_json = manager.object_names_2_accout_names(data_json)
-    return data_json
+        False
 
 
 def is_wallet_defined(logger, globals=None):
@@ -568,7 +553,7 @@ def append_account_methods_and_finish(account_object_name, account_object):
             skip_signature=0, dont_broadcast=0, forceUnique=0,
             max_cpu_usage=0, max_net_usage=0,
             ref_block=None, json=False):
-        data = _data_json(data)
+        data = manager.data_json(data)
 
         result = cleos.PushAction(
             account_object, action, data,
@@ -671,78 +656,11 @@ def append_account_methods_and_finish(account_object_name, account_object):
 
     account_object.buy_ram = types.MethodType(buy_ram, account_object)
 
-    def set_permission(
-        account_object, permission_name, authority, parent_permission_name, 
-            permission=None,
-            expiration_sec=None, 
-            skip_signature=0, 
-            dont_broadcast=0,
-            return_packed=0,
-            forceUnique=0,
-            max_cpu_usage=0,
-            max_net_usage=0,
-            ref_block=None,
-            delay_sec=0,
-            is_verbose=True,
-            json=False
-        ):
-        '''Set parameters dealing with account permissions.
 
-        - **parameters**::
-
-            permission_name: The permission to set/delete an authority for. May be
-                a string or an instance of ``eosfactory.coreinterface.Permission``.
-            authority:  None to delete; a public key string or an interface.key_arg
-                object; JSON string; a filename defining the authority.
-            parent_permission_name: The permission name of this parents permission 
-                (defaults to: "Active"). May be a string or an instance of 
-                ``eosfactory.core.interface.Permission``.
-            permission: An account and permission level to authorize, as in 
-                'account@permission'. May be an object having the attribute `name`, 
-                or a string.
-            expiration: The time in seconds before a transaction expires, 
-                defaults to 30s
-            skip_sign: Specify if unlocked wallet keys should be used to sign 
-                transaction.
-            dont_broadcast: Don't broadcast transaction to the network (just print).
-            return_packed: Used in conjunction with dont_broadcast to get the 
-                packed transaction.
-            forceUnique: Force the transaction to be unique. this will consume extra 
-                bandwidth and remove any protections against accidently issuing the 
-                same transaction multiple times.
-            max_cpu_usage: Upper limit on the milliseconds of cpu usage budget, for 
-                the execution of the transaction 
-                (defaults to 0 which means no limit).
-            max_net_usage: Upper limit on the net usage budget, in bytes, for the 
-                transaction (defaults to 0 which means no limit).
-            ref_block: The reference block num or block id used for TAPOS 
-                (Transaction as Proof-of-Stake).
-            delay_sec: Set the delay_sec seconds, defaults to 0s
-
-        '''
-        logger.TRACE('''
-        * Set account premission.
-        ''')
-
-        authority = _data_json(authority)
-
-        result = cleos.SetAccountPermission(
-            account_object, permission_name, authority, parent_permission_name,
-            permission,
-            expiration_sec, 
-            skip_signature, 
-            dont_broadcast,
-            return_packed,
-            forceUnique,
-            max_cpu_usage,
-            max_net_usage,
-            ref_block,
-            delay_sec,
-            is_verbose=False, json=True)
-        
-
-    account_object.set_permission = types.MethodType(
-                                                set_permission, account_object)
+    account_object.set_account_permission = types.MethodType(
+                                    account_set.set_account_permission, account_object)
+    account_object.set_action_permission = types.MethodType(
+                            account_set.set_action_permission, account_object)    
 
     def delegate_bw(
         account_object, stake_net_quantity, stake_cpu_quantity,
