+++ eosfactory/utils/pythonmd.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+md_file="$1"
+python3 -m eosfactory.utils.save_code "${md_file}"
